Filter alias/sym on list output when reported
When module aliases or symbolic versions are reported independently from
the loaded module they are attached to, apply pattern search also to
these alternative names.

Update modEq procedures to be able to match pattern to module name and
only its symbols or only its aliases. As pattern match is made over the
independent symbol or alias designation when reported.