Report requirement loading error on requiring module
Update the way to report requirement loading error. Failure was reported
with a "requirement not there" on requiring module message block and a
"requirement load failed" message on the top message block.

This is simplified by only reporting the "requirement load failed" error
message on the message block of the requiring module. Benefits:

* error message appears where it occurs
* error cascading is easier to understand by reading the messages from
  the main action (top message block) to the root of the error.

With this change, such error is also contained within the module
evaluation that triggers it. If this evaluation is later replaced by an
alternative requirement, all traces of this error is deleted. It was not
the case with the previous implementation where message was leaking to
the top context even if in the end the situation was solved.

Fixes #509