Modulefile man page moved from man4 to man5
modulefile man page is currently in the section 4 of man. According to
"man man", this is incorrect, as section 4 is for "Special files (usually
found in /dev)". The man page for modulefile should be in section 5,
"File formats and conventions, e.g. /etc/passwd".

This patch moves modulefile(4) to modulefile(5).

Fixes #518