Raise error when module switch in modulefile fails
Like for "module unload" and "module load", raise an error when "module
switch" command in modulefile fails to either its unload or load phase.
Error is not raised when --force option is set.

"module switch" in modulefile was previously raising error either:
* at the end of whole modulefile evaluation when conflict exists
* right at the end of the "module switch" command if switched-on module
  requirement was not satisfied.

Thus in some cases, errors during "module switch" did not lead to an
error raise and a modulefile evaluation stop. Especially when switch
unload phase fails but without registering a conflict.