Protect sh alias and function unset code
"unalias" command raises an error on "sh" kind shell when alias to unset
is not defined. Shell code is added to drop this error message and
always return success.

Same methodology is applied when unsetting function on "sh" kind shell.
An error is raised on "zsh" shell when function is not defined.
Protection code is applied for all shell of the "sh" kind for
simplicity.