Load requirement with constrained modulepath list
When auto handling mode is enabled and module requirement has to be
loaded, pass down the constrained modulepath list toward cmdModuleLoad
and getPathToModule in order to obtain resolve modulefile location
within the constrained list.

The list of modulepaths set on --modulepath prereq option is converted.
These entries are left as-is if not found in the list of enabled
modulepaths. These entries are converted into the enabled modulepaths
they match otherwise.

If required module is already loaded from a modulepath not part of
constrained list, an error is raised. This error is similar to the load
of an already loaded module with a different variant set.