Log messages at end of modulecmd processing
Create logfd and logging states to manage respectively logger pipe file
descriptor and log enablement. Logging is disabled is logger command is
an empty string or if there is no message to log.

Logs are sent during the termination phase of the modulecmd process.
Logger pipe is started for the duration of the flush operation only.