install: set value for --without opt where makes sense
Update configure script to define a value when --without-* option is
set. Only for options where it makes sense. Otherwise it falls back to
an empty string value. If this value is an empty and it makes sense
specify it to distinguish between options where --without-* makes sense
and those where it makes no sense.

Also update configure script to remove default value for --with-* option
when set without a value. Due to previous change, --with-* options now
are only treated if a value is set. Default value for installation
option applies otherwise.