Parse "not:" prefix string among advanced specifiers
Support for the "not:" prefix string is added for variant and extra
specifier search criterion.

Internal structure to represent module specification is updated to
record whether variant or extra specifier criterion is negated or not.
This boolean information is recorded next to the variant or extra
specifier name.

"not:" prefix is only accepted on module search context (module version
specification to return all matching modules).