Define stricter naming rules for variants
Instead of just requiring that first character of variant name should
correspond to the [A-Za-z0-9_] range, now variant names should only be
made of [A-Za-z0-9_-] characters. First character of these names cannot
be "-" and full name cannot be a number.

Adapt module specification parsing not to consider string parts
beginning with boolean variant prefix "+" and "~" as a variant
specification if the string after the prefix does not corresponds to a
valid variant name.

With this change, "/path/to/modul+files/foo/1.0" and "java/17.0.2+8" are
now parsed as only a module name and version rather also including a
boolean variant specification.

Fixes #527