Rework force mode mngt code in renderModulefileEvalError
Force mode is now correctly handled in unload evaluation mode. But it
still behaves specifically on other evaluation modes: message is
returned as an error and error exit code is set. This specific behavior
may be removed in the future.

Rework code to make this behavior difference visible.

Add reportInternalBugOrWarningIfForced helper proc to clarify code in
the "unload mode" branch.