Stop modfile eval if module unload fails
When a modulefile performs a "module unload" command that fails due to
dependent modules loaded, stop unload evaluation right after the failed
"module unload" command rather at the end of the modulefile evaluation
with the conflict check.

As a result later unload command in modulefile cannot solve this
conflict.

This behavior seems clearer for users as an error message was emitted
even when conflict were solved later on.

This change also corrects purge behavior when a dependent module of
currently unloading module is still loaded: unload fails right away.