Clarify test to report dependent mod loaded message
In case an unload is forced, warn about dependent modules loaded unless
for dependent reload unload phase evaluations.

The existing test to determine whether the warning should be reported or
not is clarified with this change.

As a side effect, warning message is now reported during a reload
command operation. This side effect is not compensated with a specific
test and behavior for reload command. The general idea is to limit the
number of specific behaviors.