ts: test positive isModuleEvalFailed in cmdModuleUnload
An unload evaluation may be skipped if it already failed in same
context and same evaluation tree. The skip may be demonstrated when
trying to unload twice the same super-sticky module with force mode
enabled.

This evaluation skip cannot be demonstrated with automatic module
handling mechanism.

Additional note: failed load skip may occur during restoring a broken
collection (where a module is now forbidden and another module of the
collection depends on it).