Rework handling of DepRe modules that also are UReqUn
Some *Dependent Reload* modules may also be *Useless Requirement Unload*
modules. Previously these modules were removed from DepRe list, prior
their DepRe unload processing.

Goal here is to avoid computing UReqUn module list prior DepRe and main
module evaluations. Thus DepRe modules that also are UReqUn modules are
unloaded through the DepRe unload processing.

DepRe modules that are also UReqUn are identified next to the
identification of the requirements of DepRe modules.

When processing UReqUn module unload, DepRe UReqUn modules are removed
from the DepRe list, not to run the reload processing on them. In
addition their DepRe unload evaluation is changed into a UReqUn
evaluation.

As a consequence of these changes, UReqUn modules that are DepRe are
evaluated for unload with the other DepRe modules, so prior main module
action and unload of the other UReqUn.

As a side effect of this change, DepRe modules are identified UReqUn
even if they are not a dependency of an unloaded module.