Skip reload of DepRe if requirement not satisfied
Test that DepRe module is not conflicting with loaded environment and
its requirements are loaded prior reloading it. If its requirements are
not satisfied, skip its reload.

Exception is made if DepRe module is sticky (super-sticky or sticky and
force mode disabled). In which case an attempt to reload sticky DepRe is
performed even if its requirements are not satisfied.