No force unload of DepUn of ConUn or switched-off module
When a loading module requires a Dependent Unload module, the unload of
this module was forced when it was triggered by a Conflict Unload or a
switched-off module. The unload of such Dependent Unload module is not
forced anymore by default and an error is raised if a loading module
requiring the Dependent Unload module is detected.