Use requirement modpath info to compute dependencies
Use the modulepath-specific property of the requirements to correctly
compute module dependencies. Especially to determine if a loaded or
loading module is an actual dependency of another module.

g_unmetDepHash array structure is updated to hold the specific
modulepath list next to the loaded module name.

Split isLoadedMatchSpecificPath to create isModulefileMatchSpecificPath
procedure. Useful when working on a *not-yet loaded* context.